ADJUS: Header and navBar in mobile
