+++ src/screens/patent/PatentPage.jsx
@@ -3,6 +3,7 @@ import { Typography } from "@material-tailwind/react";
 const PatentPage = () => {
   return (
     <div class="h-full mt-[80px]">
+      <Typography variant="h2">... En construcción</Typography>
       <Typography variant="h1">Nanotecnología</Typography>
       <Typography variant="h2">Tecnología TRL 6</Typography>
       <Typography variant="h6">
